# doc/api.yaml
openapi: 3.0.0
paths:
  /:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        "200":
          description: ""
  /user:
    get:
      operationId: UsersController_getAll
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
      tags: &a1
        - User
    post:
      operationId: UsersController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Body does not contain required fields
      tags: *a1
  /user/{id}:
    get:
      operationId: UsersController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Id is invalid (not uuid)
        "404":
          description: User not found
      tags: *a1
    put:
      operationId: UsersController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePasswordDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Id is invalid (not uuid)
        "403":
          description: Old password is wrong
        "404":
          description: User not found
      tags: *a1
    delete:
      operationId: UsersController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: User deleted
        "400":
          description: Id is invalid (not uuid)
        "404":
          description: User not found
      tags: *a1
  /artist:
    get:
      operationId: ArtistsController_getAll
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArtistResponse"
      tags: &a2
        - Artist
    post:
      operationId: ArtistsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArtistDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistResponse"
        "400":
          description: Body does not contain required fields
      tags: *a2
  /artist/{id}:
    get:
      operationId: ArtistsController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistResponse"
        "400":
          description: Id is invalid (not uuid)
        "404":
          description: Artist not found
      tags: *a2
    put:
      operationId: ArtistsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArtistDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistResponse"
        "400":
          description: Id is invalid (not uuid)
        "404":
          description: Artist not found
      tags: *a2
    delete:
      operationId: ArtistsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: Artist deleted
        "400":
          description: Id is invalid (not uuid)
        "404":
          description: Artist not found
      tags: *a2
  /album:
    get:
      operationId: AlbumsController_getAll
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlbumResponse"
      tags: &a3
        - Album
    post:
      operationId: AlbumsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlbumDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumResponse"
        "400":
          description: Body does not contain required fields
      tags: *a3
  /album/{id}:
    get:
      operationId: AlbumsController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumResponse"
        "400":
          description: Id is invalid (not uuid)
        "404":
          description: Album not found
      tags: *a3
    put:
      operationId: AlbumsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlbumDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumResponse"
        "400":
          description: Id is invalid (not uuid)
        "404":
          description: Album not found
      tags: *a3
    delete:
      operationId: AlbumsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: Album deleted
        "400":
          description: Id is invalid (not uuid)
        "404":
          description: Album not found
      tags: *a3
  /track:
    get:
      operationId: TracksController_getAll
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackResponse"
      tags: &a4
        - Track
    post:
      operationId: TracksController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrackDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackResponse"
        "400":
          description: Body does not contain required fields
      tags: *a4
  /track/{id}:
    get:
      operationId: TracksController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackResponse"
        "400":
          description: Id is invalid (not uuid)
        "404":
          description: Track not found
      tags: *a4
    put:
      operationId: TracksController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrackDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackResponse"
        "400":
          description: Id is invalid (not uuid)
        "404":
          description: Track not found
      tags: *a4
    delete:
      operationId: TracksController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: Track deleted
        "400":
          description: Id is invalid (not uuid)
        "404":
          description: Track not found
      tags: *a4
  /favs:
    get:
      operationId: FavoritesController_getAll
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoritesResponseModel"
      tags: &a5
        - Favorites
  /favs/track/{id}:
    post:
      operationId: FavoritesController_addTrack
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: Track added to favorites
        "400":
          description: Id is invalid (not uuid)
        "404":
          description: Track not found
        "422":
          description: Track is already favorite
      tags: *a5
    delete:
      operationId: FavoritesController_removeTrack
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Track removed from favorites
        "400":
          description: Id is invalid (not uuid)
        "404":
          description: Track not found
      tags: *a5
  /favs/album/{id}:
    post:
      operationId: FavoritesController_addAlbum
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: Album added to favorites
        "400":
          description: Id is invalid (not uuid)
        "404":
          description: Album not found
        "422":
          description: Album is already favorite
      tags: *a5
    delete:
      operationId: FavoritesController_removeAlbum
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Album removed from favorites
        "400":
          description: Id is invalid (not uuid)
        "404":
          description: Album not found
      tags: *a5
  /favs/artist/{id}:
    post:
      operationId: FavoritesController_addArtist
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: Artist added to favorites
        "400":
          description: Id is invalid (not uuid)
        "404":
          description: Artist not found
        "422":
          description: Artist is already favorite
      tags: *a5
    delete:
      operationId: FavoritesController_removeArtist
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Artist removed from favorites
        "400":
          description: Id is invalid (not uuid)
        "404":
          description: Artist not found
      tags: *a5
info:
  title: Home Library Service
  description: In-memory service for Users, Artists, Albums, Tracks and Favorites
  version: 1.0.0
  contact: {}
tags: []
servers: []
components:
  schemas:
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        login:
          type: string
        version:
          type: number
          description: Incremented on password update
        createdAt:
          type: number
          description: Creation timestamp (ms since epoch)
        updatedAt:
          type: number
          description: Update timestamp (ms since epoch)
      required:
        - id
        - login
        - version
        - createdAt
        - updatedAt
    CreateUserDto:
      type: object
      properties: {}
    UpdatePasswordDto:
      type: object
      properties: {}
    ArtistResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        grammy:
          type: boolean
      required:
        - id
        - name
        - grammy
    CreateArtistDto:
      type: object
      properties: {}
    UpdateArtistDto:
      type: object
      properties: {}
    AlbumResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        year:
          type: number
        artistId:
          type: string
          format: uuid
          nullable: true
      required:
        - id
        - name
        - year
        - artistId
    CreateAlbumDto:
      type: object
      properties: {}
    UpdateAlbumDto:
      type: object
      properties: {}
    TrackResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        artistId:
          type: string
          format: uuid
          nullable: true
        albumId:
          type: string
          format: uuid
          nullable: true
        duration:
          type: number
          description: Length in seconds
      required:
        - id
        - name
        - artistId
        - albumId
        - duration
    CreateTrackDto:
      type: object
      properties: {}
    UpdateTrackDto:
      type: object
      properties: {}
    FavoritesResponseModel:
      type: object
      properties:
        artists:
          type: array
          items:
            $ref: "#/components/schemas/ArtistResponse"
        albums:
          type: array
          items:
            $ref: "#/components/schemas/AlbumResponse"
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/TrackResponse"
      required:
        - artists
        - albums
        - tracks